+++ app/src/main/java/com/example/moodful/controller/ScreenController.kt
@@ -4,6 +4,6 @@ sealed class ScreenController(
     val route: String
 ){
     data object FrontPage : ScreenController(route = "FrontPage")
-    data object DiaryEntry : ScreenController(route = "DiaryEntry")
+    data object DiaryEntry : ScreenController(route = "DiaryEntryPage")
     data object DiaryView : ScreenController(route = "DiaryView")
 }