+++ app/src/main/java/com/example/moodful/pages/DiaryView.kt
@@ -7,7 +7,6 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import com.example.moodful.ui.theme.MoodfulTheme
 
 @Composable
@@ -26,11 +25,3 @@ fun DiaryView(modifier: Modifier = Modifier) {
     }
 }
 
-@Preview(
-    showBackground = true,
-    showSystemUi = true
-)
-@Composable
-fun DViewPreview() {
-    DiaryView()
-}