+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 3. Project files restructured
 4. Implemented basic navigation control
 5. New packages created (com.example.moodful.pages & com.example.moodful.controller)
+6. Renamed DiaryEntry.kt to DiaryEntryPage.kt
+7. First iteration of DiaryEntryPage.kt
+8. Created a separate file to preview UI elements(PreviewsPhone.kt)
+9. Refactored ScreenController.kt
 
 ### July 5, 2024